+++ skilltotal/scanners/__init__.py
@@ -22,6 +22,7 @@ from skilltotal.scanners.obfuscation import ObfuscationScanner
 from skilltotal.scanners.prompt_surface import PromptSurfaceScanner
 from skilltotal.scanners.pth_exec import PthExecScanner
 from skilltotal.scanners.python_ast import PythonAstScanner
+from skilltotal.scanners.scoped_identity import ScopedIdentityScanner
 from skilltotal.scanners.secrets import SecretsScanner
 from skilltotal.scanners.sensitive_paths import SensitivePathScanner
 from skilltotal.scanners.shell_evasion import ShellEvasionScanner
@@ -37,6 +38,7 @@ SCANNERS: list[Scanner] = [
     SensitivePathScanner(),
     NetworkScanner(),
     OAuthAuthScanner(),
+    ScopedIdentityScanner(),
     InstallScriptsScanner(),
     DynamicCodeScanner(),
     ObfuscationScanner(),